Split string utilities out into separate file
Needed to avoid circular dependency on BoutException
