+++ src/sys/makefile
@@ -5,7 +5,7 @@ SOURCEC		= array.cxx boutexception.cxx comm_group.cxx derivs.cxx \
 		  msg_stack.cxx options.cxx output.cxx \
 		  stencils.cxx utils.cxx optionsreader.cxx boutcomm.cxx \
 		  timer.cxx range.cxx petsclib.cxx expressionparser.cxx \
-	          slepclib.cxx
+	          slepclib.cxx  string_utils.cxx
 
 SOURCEH		= $(SOURCEC:%.cxx=%.hxx) globals.hxx bout_types.hxx multiostream.hxx
 TARGET		= lib
