+++ src/sys/utils.cxx
@@ -183,103 +183,3 @@ void free_cmatrix(dcomplex** m) {
   delete[] m[0];
   delete[] m;
 }
-
-
-
-/**************************************************************************
- * String routines
- **************************************************************************/
-
-// Allocate memory for a copy of given string
-char* copy_string(const char* s) {
-  char *s2;
-  int n;
-
-  if(s == NULL)
-    return NULL;
-
-  n = strlen(s);
-  s2 = (char*) malloc(n+1);
-  strcpy(s2, s);
-  return s2;
-}
-
-/// Convert a string to lower case
-const string lowercase(const string &str) {
-  string strlow(str);
-
-  std::transform(strlow.begin(), strlow.end(), strlow.begin(), ::tolower);
-  return strlow;
-}
-
-/// Convert to lowercase, except for inside strings
-const string lowercasequote(const string &str) {
-  string strlow(str);
-
-  bool quote = false, dquote = false;
-  for(string::size_type i=0;i<strlow.length(); i++) {
-    if(strlow[i] == '\'') {
-      quote ^= true;
-    }else if(strlow[i] == '"') {
-      dquote ^= true;
-    }else if( (!quote) && (!dquote) ){
-      strlow[i] = tolower(strlow[i]);
-    }
-  }
-  return strlow;
-}
-
-BoutReal stringToReal(const std::string &s) {
-  std::stringstream ss(s);
-  BoutReal val;
-  if(!(ss >> val)) {
-    throw BoutException("Could not convert string '%s' to BoutReal\n", s.c_str());
-  }
-  return val;
-}
-
-int stringToInt(const std::string &s) {
-  std::stringstream ss(s);
-  int val;
-  if(!(ss >> val)) {
-    throw BoutException("Could not convert string '%s' to int\n", s.c_str());
-  }
-  return val;
-}
-
-std::list<std::string> &strsplit(const std::string &s, char delim, std::list<std::string> &elems) {
-    std::stringstream ss(s);
-    std::string item;
-    while(std::getline(ss, item, delim)) {
-        elems.push_back(item);
-    }
-    return elems;
-}
-
-std::list<std::string> strsplit(const std::string &s, char delim) {
-    std::list<std::string> elems;
-    return strsplit(s, delim, elems);
-}
-
-// Strips leading and trailing spaces from a string
-std::string trim(const std::string &s, const std::string &c) {
-  return trimLeft(trimRight(s, c), c);
-}
-
-std::string trimRight(const std::string &s, const std::string &c) {
-  std::string str(s);
-  return str.erase(s.find_last_not_of(c)+1);
-}
-
-std::string trimLeft(const std::string &s, const std::string &c) {
-  std::string str(s);
-  return str.erase(0, s.find_first_not_of(c));
-}
-
-// Strips the comments from a string
-// This is the compliment of trimLeft
-std::string trimComments(const std::string &s, const std::string &c) {
-  std::string str(s);
-  return str.substr(0, s.find_first_of(c));
-}
-
